Added Signature Offices

+++ Options/SignatureOfficebuildingOptions.cs
@@ -5,11 +5,11 @@ public class SignatureOfficebuildingOptions
     public string Name { get; set; } = null!;
 
     public int Attractiveness { get; set; }
-    public int Wellbeing { get; set; }
-    public int Radius { get; set; }
-    public int CityOfficeEfficency { get; set; }
-    public int CityUniversityInterest { get; set; }
-    public int CityUniversityGraduation { get; set; }
+    public float Wellbeing { get; set; }
+    public float Radius { get; set; }
+    public float CityOfficeEfficency { get; set; }
+    public float CityUniversityInterest { get; set; }
+    public float CityUniversityGraduation { get; set; }
     public int Noisepollution { get; set; }
 
 }
